+++ R/mod_mc_mc.R
@@ -20,6 +20,13 @@ mod_mc_mc_ui <- function(id) {
                     collapsible = TRUE,
                     closable = FALSE,
                     width = 12,
+                    sidebar = shinydashboardPlus::boxSidebar(
+                        startOpen = FALSE,
+                        width = 80,
+                        id = ns("mc_mc_sidebar"),
+                        checkboxInput(ns("hide_lateral"), "Hide lateral genes", value = FALSE),
+                        checkboxInput(ns("hide_noisy"), "Hide noisy genes", value = FALSE)
+                    ),
                     shinycssloaders::withSpinner(
                         plotly::plotlyOutput(ns("plot_mc_mc_gene_scatter"))
                     ),