+++ R/util_genes.R
@@ -1,0 +1,10 @@
+annotate_genes <- function(genes, dataset) {
+    lateral_genes <- get_mc_data(dataset, "lateral_genes")
+    noisy_genes <- get_mc_data(dataset, "noisy_genes")
+
+    case_when(
+        genes %in% lateral_genes ~ "lateral",
+        genes %in% noisy_genes ~ "noisy",
+        TRUE ~ "other"
+    )
+}