+++ NEWS.md
@@ -8,6 +8,7 @@
 * Fix: differential expression between metacells did not work with most recent version of `qs`.
 * Added Lasso selection in Annotation and Differential Expression tabs.
 * Added "# of metacells with significant inner-fold" to the "QC" tab.
+* Filter lateral/noisy genes from diff expression plots.
 
 # MCView 0.2.9